app/requirements.txt
pytesseract
opencv-python
pillow